+++ src/lib/CoinbasePro/Authenticated/Deposit.hs
@@ -4,6 +4,9 @@
 
 module CoinbasePro.Authenticated.Deposit
     ( Deposit (..)
+    , DepositRequest (..)
+    , DepositResponse (..)
+    , PaymentMethodId
     ) where
 
 import           Data.Aeson                         (FromJSON, ToJSON,
@@ -12,9 +15,10 @@ import           Data.Aeson                         (FromJSON, ToJSON,
 import           Data.Aeson.Casing                  (snakeCase)
 import           Data.Aeson.TH                      (defaultOptions, deriveJSON,
                                                      fieldLabelModifier)
-import           Data.Text                          (Text)
+import           Data.Text                          (Text, pack, toLower)
 import           Data.Time.Clock                    (UTCTime)
 import           Data.UUID                          (UUID)
+import           Servant.API
 
 import           CoinbasePro.Authenticated.Accounts (AccountId)
 import           CoinbasePro.Types                  (CreatedAt, UserId)
@@ -33,12 +37,28 @@ data DepositDetails = DepositDetails
 deriveJSON defaultOptions { fieldLabelModifier = snakeCase } ''DepositDetails
 
 
+newtype PaymentMethodId = PaymentMethodId UUID
+  deriving (Eq)
+
+
+instance Show PaymentMethodId where
+  show (PaymentMethodId p) = show p
+
+
+deriveJSON defaultOptions { fieldLabelModifier = snakeCase } ''PaymentMethodId
+
+
+instance ToHttpApiData PaymentMethodId where
+    toUrlPiece   = toLower . pack . show
+    toQueryParam = toLower . pack . show
+
+
 newtype UserNonce = UserNonce Int
   deriving (Eq, Show, ToJSON, FromJSON)
 
 
 data Deposit = Deposit
-    { dId          :: UUID
+    { dId          :: PaymentMethodId
     , dType        :: Text
     , dCreatedAt   :: CreatedAt
     , dCompletedAt :: UTCTime
@@ -46,7 +66,7 @@ data Deposit = Deposit
     , dProcessedAt :: Maybe UTCTime
     , dAccountId   :: AccountId
     , dUserId      :: UserId
-    , dUserNonce   :: Maybe Int
+    , dUserNonce   :: Maybe UserNonce
     , dAmount      :: Double
     , dDetails     :: Maybe DepositDetails
     } deriving (Eq, Show)
@@ -65,3 +85,24 @@ instance FromJSON Deposit where
     <*> (o .:? "user_nonce")
     <*> (read <$> o .: "amount")
     <*> (o .:? "deposit_details")
+
+
+data DepositRequest = DepositRequest
+    { amount          :: Double
+    , currency        :: Text
+    , paymentMethodId :: PaymentMethodId
+    } deriving (Eq, Show)
+
+
+deriveJSON defaultOptions { fieldLabelModifier = snakeCase } ''DepositRequest
+
+
+data DepositResponse = DepositResponse
+    { rId       :: UUID
+    , rAmount   :: Double
+    , rCurrency :: Text
+    , rPayoutAt :: UTCTime
+    } deriving (Eq, Show)
+
+
+deriveJSON defaultOptions { fieldLabelModifier = snakeCase . drop 1 } ''DepositResponse